requirements.txt
Django==2.2
django-crontab==0.7.1
Jinja2==2.10.1
mysqlclient==1.4.6
sqlparse==0.3.0
tqdm==4.32.2
xlrd==1.2.0
Pillow==7.2.0
itsdangerous==1.1.0
PyMySQL==0.10.0
PyYAML==5.3.1

